PY-105 Split filters between alpha.filters and internal.filters
This commit introduces a distinction between high-level and internal
filters. This will allow us to modify the API in the future when we
introduce a new API version - `v1` alongside the current `alpha`.

High-level are kept in `neptune_fetcher.alpha.filters` and are
meant to be constructed by the user. (Their API is meant to be fully
compatible with the previous version).

Low-level/internal filters are put to `neptune_fetcher.internal.filters`
and their class names start with underscore, i.e.: `_Filer`,
`_Attribute`, etc.

User-accessible functions need to convert user-supplied filters to the
ones expected by internal API and this happens in the `_resolve_*`
functions in `neptune_fetcher.alpha._internal` using `_to_internal`
methods on the high-level objects.